Landing Site to give a bit about my HTML/CSS/Javascript skills
This is the main page that is my main source for contact and this is the best way to reach me anytime. This is best used as the main hub for any information about me because it lets me be creative and focused on what matters which is clean and simple code.